requirements.txt
# Core Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# File Type Detection
python-magic>=0.4.27

# PDF Processing (PyMuPDF replaces pdfplumber + pdf2image)
PyMuPDF>=1.24.0
pypdf>=3.17.0

# OCR (Tesseract — fast, accurate Arabic/multilingual support)
pytesseract>=0.3.10
Pillow>=10.4.0

# Image Preprocessing
opencv-python-headless>=4.9.0

# Encoding Detection
charset-normalizer>=3.3.0

# Office Documents
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.2.0
lxml>=4.9.0

# Language Detection
langdetect>=1.0.9

# Database
sqlalchemy>=2.0.25

# Utilities
aiofiles>=23.2.1
pydantic>=2.5.3
numpy>=1.24.0

# RAG Pipeline
sentence-transformers>=2.3.1
faiss-cpu>=1.9.0
langchain>=0.1.6
langchain-text-splitters>=0.0.1
tiktoken>=0.5.2

# Additional Format Support
beautifulsoup4>=4.12.3
striprtf>=0.0.26
